por temas de compatibilidad de la base de datos donde el id ciudadano es biginit, en el backend se debe de cambiar su atributo de int a long, eos se hizo donde fuera necesario

+++ Backend/Delta/src/main/java/com/sistema_de_vacunacion/Delta/Vacunacion/dto/RegistrarVacunacionDTO.java
@@ -9,7 +9,7 @@ import lombok.Data;
 public class RegistrarVacunacionDTO {
 
     @NotNull(message = "El ID del ciudadano es obligatorio")
-    private Integer idCiudadano;
+    private Long idCiudadano;
 
     @NotNull(message = "El ID de la vacuna es obligatorio")
     private Integer idVacuna;
