+++ Backend/Delta/src/main/java/com/sistema_de_vacunacion/Delta/usuario/AdministradorRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 
-public interface AdministradorRepository extends JpaRepository<Administrador, Integer> {
+public interface AdministradorRepository extends JpaRepository<Administrador, Long> {
 
     // Buscar administrador por correo para autenticación
     Optional<Administrador> findByEmail(String email);
